app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
psycopg2-binary==2.9.10
redis==4.0.2
prometheus-flask-exporter==0.18.7
prometheus-client==0.13.1